demos/securecoding/03-python-dependencies/fixed/requirements.txt
#
# This file was generated by pip-compile --generate-hashes
# by the following command:
#
#    pip-compile --generate-hashes --output-file requirements.txt requirements.in
#
# To regenerate, run:
#    pip-compile --generate-hashes --output-file requirements.txt requirements.in
#
# NOTE: The hashes below are ILLUSTRATIVE PLACEHOLDERS for this training demo.
#       Real hashes must be generated by running pip-compile on your own machine.
#       Never commit manually crafted hashes to a production lock file.
#       Run: pip-compile --generate-hashes requirements.in
#
# FIX: Every package (direct AND transitive) is pinned to an exact version.
#      SHA-256 hashes verify the downloaded wheel/sdist has not been tampered with.
#      Installing with: pip install --require-hashes -r requirements.txt
#      will reject any package whose hash does not match — preventing
#      dependency confusion, typosquatting, and supply-chain substitution attacks.

blinker==1.8.2 \
    --hash=sha256:8f77b09d3bf7c795e969e9486f39c2c5e9c39d4ed07dd26ebcd8df8a9827fce8 \
    --hash=sha256:1779309f71bf239144b9399d06ae925637cf6634cf6bd131104184531bf67c01
    # via flask

click==8.1.7 \
    --hash=sha256:ae74fb96c20a0277a1d615f1e4d73c8414f5a98db8b799a7931d1582f3390c28 \
    --hash=sha256:ca9853ad459e787e2192211578cc907e7594e294c7ccc834310722b41b9ca6de
    # via flask

flask==3.0.3 \
    --hash=sha256:34e815dfaa43c9d9e670b7f56a945da3b02cc73f3bbb35ea7d696f0e5b6e4e7c \
    --hash=sha256:ceb27b0af3823ea2737928a4d99d125a06175b8dbe2c3b716b9f29e2930f87a1
    # via -r requirements.in

itsdangerous==2.2.0 \
    --hash=sha256:e0050c0b7da1eea53ffaf149c0cf592038c3b5a9e69c5bff969e45fb0c9ca31b \
    --hash=sha256:c6242fc49e35958c8b15141343aa660db5fc54d4f13a1db01a3f5891b98700ef
    # via flask

jinja2==3.1.4 \
    --hash=sha256:4a3aee7acbbe7303aede8e9648d13b8bf88a429282aa6122a993f0ac800cb369 \
    --hash=sha256:bc5dd2abb727a5319567b7a813e6a2e7318c39f4f487cfe6c89c6f9c7d25197d
    # via flask

markupsafe==2.1.5 \
    --hash=sha256:d283d37a890ba4c1ae73ffadf8046435c76e7bc2247bbb63c00bd1a709c6544b \
    --hash=sha256:2174c595a0d73a3080ca3257b40096db99799265e1c27cc5a610743acd86d62f
    # via jinja2, werkzeug

# FIX: Pillow pinned to 10.3.0 — NOT the vulnerable 9.0.0 (CVE-2022-22817).
#      pip-audit will confirm no known CVEs for this version.
pillow==10.3.0 \
    --hash=sha256:2697b21f4de8d0802e6869e7c7f423ca28f08e3e3fb4eba22b82e7d0e73e4a5b \
    --hash=sha256:8b7c84fca48a11a30c2c8ffece7b39949a88db0d0c9a0fce539614b3ad1f0040
    # via -r requirements.in

pyyaml==6.0.1 \
    --hash=sha256:bfdf460b1736c775f2ba9f6a92bca30bc2095067b8a9d77876d1fad6cc3b4a43 \
    --hash=sha256:0d5abf9c72b08e723b53c7d7c560c28d22fc0bbf38e78cee2f6f1f87e50e6dfe
    # via -r requirements.in

requests==2.31.0 \
    --hash=sha256:58cd2187423d77b898475a6ff50f0b62a5a8e0e82ab2a50fbdb21f69f3b4b3c6 \
    --hash=sha256:942c5a758f98d790eaed1a29cb6eefc7ffb0d1cf7af05c3d2791656dbd6ad1e1
    # via -r requirements.in

werkzeug==3.0.3 \
    --hash=sha256:fc9645dc43e03e4d630d23143a04a7f947a9a3b5727cd535fdfe155a17cc48c8 \
    --hash=sha256:097c403f0a9083a5dc3db839c4c7abfe7bd5c0018c7c1f5f0c0e0a49ccca9ec7
    # via flask
